Backend | CB-222 Create special test utils for public events
