+++ backend/libs/public-messages/src/index.ts
@@ -5,3 +5,5 @@ export * from './inviting-applicants/command/applicant-invitation.command';
 export * from './inviting-applicants/event/applicant-invitation.public-event';
 export * from './shared/command/external-command';
 export * from './shared/event/public-event';
+export * from './shared/test.utils';
+
