+++ backend/src/write-side/inviting-applicants/domain/applicant-invitation-id.valueobject.ts
@@ -2,6 +2,8 @@ import uuid = require('uuid');
 import { AggregateId } from '../../shared-kernel/domain/aggregate-id.valueobject';
 
 export class ApplicantInvitationId implements AggregateId {
+  private readonly TYPE = 'ApplicantInvitationId';
+
   private constructor(readonly raw: string) {}
 
   static generate() {
