+++ backend/src/write-side/shared-kernel/domain/aggregate-root.spec.ts
@@ -1,0 +1,5 @@
+describe("Feature: Aggregate Root", () => {
+
+
+
+})
